+++ src/main/java/com/capg/ftb/dao/IBookingDAO.java
@@ -4,12 +4,14 @@
 
 package com.capg.ftb.dao;
 
+import java.math.BigInteger;
+
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 import com.capg.ftb.model.Booking;
 
 @Repository
-public interface IBookingDAO extends JpaRepository<Booking,Integer>{
+public interface IBookingDAO extends JpaRepository<Booking,BigInteger>{
 
 }
