+++ src/main/java/com/capg/ftb/service/IFlightService.java
@@ -4,6 +4,7 @@
 package com.capg.ftb.service;
 
 
+import java.math.BigInteger;
 import java.util.List;
 
 import com.capg.ftb.model.Flight;
@@ -12,12 +13,12 @@ public interface IFlightService {
 
 	public Flight addFlight(Flight flight);
 	
-	public Flight viewFlight(int flightNumber); // throws FlightNotFoundException;
+	public Flight viewFlight(BigInteger flightNumber); // throws FlightNotFoundException;
 	
 	public List<Flight> viewAllFlights();
 	
-	public Flight deleteFlight(int flightNumber);
+	public Flight deleteFlight(BigInteger flightNumber);
 	
-	public Flight modifyFlight(int flightNumber);
+	public Flight modifyFlight(BigInteger flightNumber);
 	
 }
